pyproject.toml
[project]
name = "data-ingestion-service"
version = "0.1.0"
description = "Scalable data ingestion service with MinIO storage"
requires-python = ">=3.9"
authors = [{name = "Data Team"}]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "minio==7.2.0",
    "boto3==1.29.7",
    "kaggle==1.5.13",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "scrapy==2.11.0",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "apscheduler==3.10.4",
    "sqlalchemy==2.0.23",
    "python-multipart==0.0.6",
    "httpx==0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
